Очищение старых сконкатенированных файлов после комплияции

+++ tools/rmdir.js
@@ -14,6 +14,8 @@ try{
             console.info(`\x1b[31m${file}\x1b[37m`)
         },
         (dir) => {
+            if(dir === path)
+                return
             fs.rmdirSync(dir)
             console.info(`\x1b[31m${dir}\x1b[37m`)
         });
